Nouveau code C
Code C qui a vocation d'être testé sur la carte Linux (pas encore final)

+++ C/utils.h
@@ -6,4 +6,4 @@
 
 
 int Recuperateur(const char *chemin, char *buffer, size_t bufferSize);
-uint8_t convertisseur32bits(const char *binaire);
+uint32_t convertisseur32bits(const char *binaire);
